framework/sunflower: Allow powering on from dead battery
Since the Sunflower main battery supports RTC function, the battery
present pin will be changed to the RTC pin of the LDO 3.3V output.

The battery present pin will be high if the battery is drained. Follow
the 16-inch project to read the battery register for 30 seconds to check
the battery, when the battery present pin is high during EC initial.

BRANCH=fwk-sunflower-26784
BUG=app.clickup.com/t/86eqf6heh
TEST=EC can boot from drained battery when AC is connected. And drained battery can be charged on sunflower

Signed-off-by: Josh-Tsai <josh_tsai@compal.com>