brox: Allow PR_Swap to source in VIF
Currently the Brox VIF states that we do not PR_Swap as a sink. But,
should_swap_to_source() in the PDC power manger defines conditions where
Brox would initiate a PR_Swap to switch to a source. This causes
COMMON.PROC.PD.5 to fail because it sees a PR_Swap which contradicts the
Requests_PR_Swap_As_Snk field in the VIF.

This updates the Brox VIF to tell the tester that Brox may PR_Swap from
sink to source.

BUG=b:373507684
TEST=None

Change-Id: I41f42a411701d993b7092dab2e3cd9d8b9d440ee
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/ec/+/5958987
Commit-Queue: Tristan Honscheid <honscheid@google.com>
Reviewed-by: Keith Short <keithshort@chromium.org>
Tested-by: Tristan Honscheid <honscheid@google.com>
Code-Coverage: Zoss <zoss-cl-coverage@prod.google.com>