Revert "fwk: sunflower: disable ps2mouse function"
Explanation:

EVT and DVT1 schematic shows that touchpad I2C is connected to the EC
(and SoC).
But the resistors to the EC were not actually in the BOM and did not get
populated during SMT.
This means for DVT1 the schematic is still the same but BOM is updated
and the resistors (RK920, RK921) are placed during DVT1 SMT.

This change is necessary because Windows and the installer still does not
have inbox drivers for the Intel I2C controller, even in 24H2.
So touchpad does not work in the windows installer, nor a clean windows
installation.

This reverts commit e8e08b5513a28d43872464737748ef5527b63855.

Enable PS/2 mouse function for DVT.

BRUNCH=fwk-sunflower-26784
BUG=https://app.clickup.com/t/86eq8r2vk
TEST=ps2 mouse is workable on EVT rework DUT(rework TP I2C connect to EC)