fwk: sunflower: Ignore lid switch for power button press
Currently the EC ignores power button presses when the lid is closed.
But when the user connects an external monitor and closes the lid,
the user may press the power button to power on/off or wake/suspend the system.
Add the config for ignoring the lid switch when the power button is pressed.

BRANCH=fwk-sunflower-26784
BUG=NONE
TEST=Close the lid and the system can power on when the power button is pressed.

Signed-off-by: LeilaCY-Lin <LeilaCY_Lin@compal.com>