fwk: sunflower: workaround: Notify SCI event using GPIO
When all drivers installed, system sometimes not dealing SCI event in S0i3.
This change use the GPIO to notify the SCI event.
This is necessary to make restricted standby work.
Intel is debugging eSPI SCI events, but has not found a solution, so we
need to switch to GPIO right now as a workaround.

This requires a new GPIO in DVT1. It is not connected in EVT.
The result is that using this on EVT mainboards will break SCI events.

BRANCH=fwk-sunflower-26784
BUG=https://app.clickup.com/t/86eqbbeed
TEST=zmake build sunflower pass
TEST=Windows `powercfg /spr` shows 'Processes Disabled in Standby' 1,
which means restricted standby was enabled.

Signed-off-by: Josh-Tsai <josh_tsai@compal.com>