fwk: don't call the battery and charger function in PD code if no support
Some of the PD functions are related to battery and charger. If EC does not support
battery and charger, EC should not call those functions in the code to avoid the
compiler error.

This change uses the #ifdef and IS_ENABLE() to avoid the function being called if EC
does not support it.

BRANCH=fwk-main
BUG=None
TEST=PD function works well on the marigold unit TEST=zmake build marigold pass

Signed-off-by: Josh Tsai <Josh_Tsai@compal.com>