fwk: sunflower: enable power LED when S5S3 chipset startup
For now, the Sunflower project only enables power LED when power
state G3S5(for illuminating immediately when the user presses the
power button #1157) and disables when power state S3S5. If the
system is powered up from S5, the power LED will not be enabled.
So, we also need to enable the power LED when the power state S5S3
chipset startup to turn on the power LED correctly.

This commit fixes a bug in #1157.

BRANCH=fwk-sunflower-26784
BUG=NONE
TEST=Press the power button as soon as the system shutdown and the
power LED turns off(power state in S5), and check that the power
LED is on.

Signed-off-by: LeilaCY-Lin <LeilaCY_Lin@compal.com>