fwk/sunflower: trigger input cover diag if no KB controller
Make the diagnostics more accurate - if BOARD_ID is detected but keyboard
controller does not respond, it should also trigger the diagnostic.
The interface isn't a solid connector like on 13in or 16in but pogo pins
instead, so it might happen that some of them make contact (e.g.
BOARD_ID) but others do not.
Diagnostic should be triggered if there is anything wrong with the
connection.

Rename diag to INPUT_COVER because it doesn't just detect the touchpad
but the whole input cover.

BUG=Implementing UEFI inputmodule app
BRANCH=fwk-sunflower-26784
TEST=Tape over touchpad I2C pins on JTP1 and check whether that results
in diagnostics being triggered (diag lights blinking)

Signed-off-by: Daniel Schaefer <dhs@frame.work>