fwk: disable TP interrupt before reenable TP
When reenable TP, EC will reset TP. In the reset process, the
device will assert the interrupt to indicate that it has been
initialized but this interrupt will that EC think there is an
interrupt event so stuck in the loop. So we need to disable the
interrupt before reenabling the TP PS/2 mouse function.

BRANCH=fwk-main
BUG=https://app.clickup.com/t/86er0699h
TEST=Disable TP HID driver, PS/2 mouse function can work without
 1 min delay.

Signed-off-by: LeilaCY-Lin <LeilaCY_Lin@compal.com>