tulip: add is_throttling gpio func to gpu
add new function definition to GPU descriptor definition.
Also increment supported version to 2. As this is a breaking change
to the GPIO configuration definition. so cannot be backwards compatible
with older released BIOS versions.

This is a WIP, as the full impelemntation for added features is not
implemented yet.

Signed-off-by: Kieran Levin <ktl@frame.work>
(cherry picked from commit 2135d74129535c41a2c31f57ab791027fac8cfdd)