fwk: handle extended control message EPR KeepAlive
Workaround to handle the message type is EPR KeepAlive. Note on
the new PD firmware, it will pass the EPR keepalive messages to
the EC right now. To avoid spamming the ec log until we figure
out how to disable this.

BRANCH=fwk-main
BUG=https://app.clickup.com/t/86er1uac7
Test=Used PD firmware(lotus_3.7.0_test) on lotus and checked the
message of EPR KeepAlive message not show.

Signed-off-by: LeilaCY-Lin <LeilaCY_Lin@compal.com>