fwk: lilac: do not mask UCSI_CMD_GET_ALTERNATE_MODE
Have the EC allow PD to get alternate modes from the host
for the port partner.

This removes a work around in the EC, where alternate mode requests to
SOP' or SOP " would be rewriteen to go to the connector. Which is
actually not valid.

This seemed to cause issues on UCSI initialization. Where the UCSI init
sequence would get an error response to port altnernate modes for the
connector/cable. And then would erroniously display device mismatch
warnings

BUG=https://app.clickup.com/t/86equ702f
TEST=run HLK test on 24H2, and we see that GET_ALT_MODE passes.
Reboot and attach a DP dongle and ensure that no connector mismatch
notifications show up in Windows 24H2.
Then reboot, and try again.

BRANCH=fwk-lilac-27116

Signed-off-by: Kieran Levin <ktl@frame.work>