fwk: reset touchpad when in PTP mode
When the C cover is removed and reattached, the touchpad's report
mode will initial to PTP mode we need to reset the touchpad to
mouse mode so that EC can recognize the report mode communication.

BRANCH=fwk-main
BUG=https://app.clickup.com/t/86er7a9bu
TEST=Check that HID I2C shows a yellow bang in DM, remove the C
cover, and enter the S0ix state. Wake up the system and reattach
the C cover back to check that the touchpad function works normally.

Signed-off-by: LeilaCY-Lin <LeilaCY_Lin@compal.com>