fwk: dogwood: schematic change EC_H_PROCHOT_L
The EC_H_PROCHOT_L signal on the DVT1 sets to GPIOA1. This pin shoudn't
be used when EC controls the SPIP mode.

We change the EC_H_PROCHOT_L signal to GPIO67 on the DVT2 mainboard.
This change use the board_get_version() to control the prochot pin
on different mainboard.

BRANCH=fwk-dogwood-27111
BUG=None
TEST=increase temperature to trigger thermal high, check EC can
     assert the prochot pin correctly.

Signed-off-by: Josh Tsai <Josh_Tsai@compal.com>