fwk: support ac boot feature for mini-pc
The laptop designs that if the power on AC is enabled EC checks the
extpower gpio status to decide to power on the system or not.

In the commit a21354caecf2bfa142be63e276251afb581036a9, we disable
the power on AC feature if CONFIG_PLATFORM_EC_EXTPOWER_GPIO does
not be enabled.

This change modifies the logic to support the power on AC feature
without detecting the extpower gpio status for Mini-PC

BRANCH=fwk-main
BUG=https://app.clickup.com/t/86ermuw98
TEST=use ec console "memmap set 0x148 0x01" to set the power on ac flag
     check the EC can power on the system when the PSU power on.

Signed-off-by: Josh Tsai <Josh_Tsai@compal.com>