fwk: sunflower: modified the ISL9241 fast discharge logic
When the active charge port is removed, EC needs to enable the ISL9241
bit 15 to fast discharge the B+ voltage to avoid to detect the wrong
AC present status.After EC enables the fast discharge, the B+ voltage
needs to spend about 10 ms to reduce the voltage from 4.5V to 3V.

This change modified the logic to ensure the fast discharge can be
always trigger when the active charge port is removed. And wait
CONFIG_DISCHARGE_TIMING to ensure the B+ is less than 3V and then
disable the fast charge feature to avoid the leakage current.

BRANCH=fwk-sunflower-26784
BUG=https://app.clickup.com/t/86ern1ypw
TEST=Measure the waveform, the acok signal will not assert when user
     remove the type-c adapter.
TEST=use the console command "i2ctrace" to check the register is
     controlled as the expected

Signed-off-by: Josh Tsai <Josh_Tsai@compal.com>