fwk: KB auto brightness just one low level
You only need KB backlight if the environment is extremely dim and your
keyboard is not illuminated.
Even sometimes when you're in a fully dark room and you turn up your
display brightness, if reflects off the environment and lights up the
keyboard enough to see it.

BUG=Backlight brightness sometimes toggles back and forth because
there's no hysteresis. It doesn't need that many levels. And it doesn't
need to be on when there's environmental light.
BRANCH=fwk-main
TEST=FN+SPACE to toggle kb backlight brightness to auto. You'll see
backlight off, unless the environment is very dark (so dark you can't
see the keycap labels)

Signed-off-by: Daniel Schaefer <dhs@frame.work>