fwk/tulip: move the ccg8s init into cypress code
this moves the ccg8s initialization code out
of the gpu into the pd files.

This also cleans up the shutdown/deinit of the
ccg8 pd controller when the fxbeam is opened.
And removes the power off function in the gpu
as the PD controller will be left on whenever
the EC is powered on.

BUG=https://app.clickup.com/t/86erqbffx
BRANCH=fwk-main
TEST= flash on tulip, and ensure the
port functions of the rear port can work correctly.
Eg sink path, charging, etc.

TEST=use cypdstatus 2 to ensure the controller
is recognized correctly.

Signed-off-by: Kieran Levin <ktl@frame.work>