fwk: led: runtime to update the led duration
In the commit a30cccd442135ce8517757468c6073e8264bf66e, we only modified the
LED driver to reduce the tick from 200ms to 10ms when the system enter suspend.

We did not modify the LED driver to change the duration at runtime.
It causes the period-ms value not to meet the actual behavior.

This change modified the LED driver to change the color duration
at runtime.

BRANCH=fwk-main
BUG=None
TEST=When the system enter suspend, the led breathing can meet
     the period-ms value.

Signed-off-by: Josh Tsai <Josh_Tsai@compal.com>