fwk: dogwood: convert the temperature ratio to fan duty
In the commit 6ec3388ccfba909c1c681cc8738bed94f580f63a, I used the
temperature ratio to be the fan duty. But when the user enables the
fan always on configuration, the fan target duty will keep in minimum
duty even if the temperature is over the minimum temperature.

In this change, EC will convert the temperature ratio to fan duty and
start the duty from minimum duty instead of the 0.

BRANCH=fwk-dogwood-27111
BUG=None
TEST=Monitor the temperature and target duty; the target duty's value
     is expected based on the temperature ratio.
     e.g. configure the max temp 100 and min temp 60, the target duty
     increases ~2% per Celsius

Signed-off-by: Josh Tsai <Josh_Tsai@compal.com>