fwk: dogwood: add more cases in update_os_power_slider()
The switch statment in update_os_power_slider is not exhaustive.
Linux seems to use the EC_DC_* settings for some reason (and realistically
we don't want to wait for a fix there, even if they should use AC*)
Windows also wants to sometimes set it to EC_AC_BEST_EFFICIENCY.

Previously, in all of these cases, the default would execute,
which just means leaving it on the previous setting.

Now the switch statment selects more of the options (correctly too), and
default: is just the lower setting.

BUG=https://app.clickup.com/t/86erte3a0
TEST=On windows: switch into low-power, you should see balanced settings,
not the previous setting (which might be performance)
TEST=On linux, change /sys/firmware/acpi/platform_profile to all settings
presented in /sys/firmware/acpi/platform_profile_choices, you should see
the EC console react (~115W for the first 2 or ~140W for the performance one),
and see the new power values in `sensors`
BRANCH=fwk-dogwood-27111