fwk: dogwood: improve the 5VSB enter/exit flow
If EC calls the exit 5VSB flow, it needs to wait for the pok signal to assert.
According to the waveform, PSU needs to spend about 400 - 420 ms to assert the
pok signal.

This change creates the hook deferred function to re-check the power monitor
and 5VSB status to avoid EC missing the power monitor change the event while
it is waiting for the pok signal.

BRANCH=fwk-dogwood-27111
BUG=https://app.clickup.com/t/86et6gp27
TEST=Reduce the current to 500mA, plugin the USBA device to let ina236
     power monitor asserted the interrupt pin to simulate the OCP.
     1.Plugin the USB-A device to trigger OCP and then unplug it quickly.
       EC exits the 5VSB first and then enter the 5VSB correctly.
       [1393.059200 power 5vsb exit]
       [1393.657700 power state 3 = S3, in 0x0005]
       [1393.661500 power 5vsb enter]
       [1393.736100 power state 3 = S3, in 0x0004]

Signed-off-by: Josh Tsai <Josh_Tsai@compal.com>