fwk: add the temp sensor id off
We would like to use the temp sensor id to link the sensor source
for customized fan control feature. But we need a sensor ID (e.g. off)
to wait for the system to update the sensor source.

This change adds the TEMP_SENSOR_OFF id in the enum temp_sensor_id.

BRANCH=fwk-dogwood-27111
BUG=None
TEST=zmake build marigold pass
(can't test this change,does not affect any function)

Signed-off-by: Josh-Tsai <josh_tsai@compal.com>