common/flash: Fix flash protect v2 race condition
There is a race condition in `flash_command_protect_v2`. The way the API
is intended to be used is to first issue a `FLASH_PROTECT_ASYNC` action
that contains the flags that we want to set. We then poll using the
`FLASH_PROTECT_GET_RESULT` action. If the `FLASH_PROTECT_GET_RESULT`
action happens before the deferred or worse, during the deferred then
the mask and flags are wiped out.

This change makes it so the mask and flags are latched when the
`FLASH_PROTECT_ASYNC` action is received.

BUG=b:412723132
TEST=Verify ectool flashprotect works on AL device.
LOW_COVERAGE_REASON=No existing tests for this command.

Change-Id: Ie9c359da4698feb0c8bd977ccc51d0d2ea4fddb9
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/ec/+/6520620
Code-Coverage: Raul Rangel <rrangel@chromium.org>
Reviewed-by: Tristan Honscheid <honscheid@google.com>
Tested-by: Raul Rangel <rrangel@chromium.org>
Auto-Submit: Raul Rangel <rrangel@chromium.org>
Commit-Queue: Raul Rangel <rrangel@chromium.org>