fwk:sunflower:Disable the power button LED when power state S5G3
Sunflower will immediately enable the power button LED when EC
registers the power-on event(power state G3S5)
[PR1157](https://github.com/FrameworkComputer/ec/pull/1157).
But we only disable the power button LED when the power state
is S3S5.

Disable the power button LED when the power state is S5G3 to make
sure the LED is disabled.

BRANCH=fwk-sunflower-26784
BUG=https://app.clickup.com/t/86er07k6n
TEST=Check the power button LED is disable when power-up S5 sequence
time out back to G3 state.
TEST=To force this, you can lower ap_boot_delay in sunflower/src/power_sequence.c

[7445.476900 power state 0 = G3, in 0x0005]
[7445.478100 power state 6 = G3->S5, in 0x0005]
[7445.714200 setup power s5 up!]
[7445.715300 power state 1 = S5, in 0x0065]
[7455.718900 timeout waiting for S5 exit]
[7455.720400 power state 11 = S5->G3, in 0x0065]
[7455.732600 power state 0 = G3, in 0x0005]

Signed-off-by: LeilaCY-Lin <LeilaCY_Lin@compal.com>