fwk: check GPU PD chip is ready in battery response function
On tulip, the number of PD chips is three.
If gpu didn't install, we will also communicate with the GPU PD.
EC needs to avoid to communicate with the PD chip that is not power.
So check GPU PD chip is ready.

BRANCH=fwk-main
BUG=I found when checking PR1528 that tulip/lotus would write invalid
GPU PD port2, and whether the GPU PD is ready.
TEST=Not communicate with GPU PD unit the GPU PD is ready

Signed-off-by: johnwc_yeh <JohnWC_Yeh@compal.com>