fwk: Immediately adjust KB backlight if switching to auto
last_kbbl_led_brightness should be an int compared with kb_brightness,
which is int, so it does not make as a bool.

BUG=When cycling through brightness with FN+SPACE after high, switching
to auto the backlight will stay high. It doesn't adjust to the current
brightness until the ALS brightness changes between off/on.
TEST=press FN+space 4 times from off and make sure that in auto mode
kb backlight immediately updates to ultra-low or off
BRANCH=fwk-main

Signed-off-by: Daniel Schaefer <dhs@frame.work>