fwk: cypd: call pd task event to update the battery information
EC spends about 150ms to update the battery information, which causes
the GPU to configure too late to turn on the GPU power.

This change uses the PD task event to update the battery information
after EC powers on 500ms to wait the gpu configuration to complete.

BRANCH=fwk-main
BUG=https://app.clickup.com/t/86etf8451
TEST=use ec console to check the timestamp of the gpu configure complete
and close the interposer door when the battery is fully charged, and it
will update the battery message to the GPU PD.

[Before]
[0.478900 PD2 Reset Complete]
[0.491600 Controller 2  App1 version B:3.7.0.197 , AP:0.0.D]
[0.495600 Controller 2  App2 version B:3.7.0.197 , AP:0.0.D]
[0.505100 GPU descriptor read complete]

[After]
[0.340000 Controller 2  App1 version B:3.7.0.197 , AP:0.0.D]
[0.345200 Controller 2  App2 version B:3.7.0.197 , AP:0.0.D]
[0.360100 GPU descriptor read complete]
[0.362800 power state 3 = S3, in 0x0007]

Signed-off-by: Josh Tsai <Josh_Tsai@compal.com>