fwk: dogwood: don't enter 5vsb when force to enable PSU
In this commit 871089f0945fb91d623ff6ad872f66dd800e358e, I missed the
condition "force_enable_psu" to call power_5vsb_enter() and caused the
system to drop the power when the system enters suspend with force_enable_psu.

This change adds the logic to the power_5vsb_enter() function.
When the force_enable_psu flag is set, just return the function and do
not run the turn-off sequence.

BRANCH=fwk-dogwood-27111
BUG=https://app.clickup.com/t/86etpj066
TEST=Enable the "force_enable_psu", system can enter/exit from standby mode.
     The PSU does not be turned off.
TEST=disable the "force_enable_psu", system can enter/exit from standby mode.
     The PSU is turned off.

Signed-off-by: Josh Tsai <Josh_Tsai@compal.com>