fwk: minipc-diagnostic: don't run the diagnostic if no HW/SW error
EC will turn off the PSU power when the system shutdown or enters standby mode.
In this case, EC calls the diagnostic_power_glitch_detect() and run the diagnostic.

EC shouldn't execute the diagnostic if there is no HW or SW error, so EC needs to
clear the variable "run diagnostics" to ensure diagnostic is stopped

BRANCH=fwk-dogwood-27111
BUG=diagnostic will be executed if the system enters standby mode
BUG=diagnostic will be executed if the system does the normal shutdown
TEST=EC does not print the diagnostic log if the system enters standby mode
     or does the normal shutdown after 90s.

Signed-off-by: Josh Tsai <Josh_Tsai@compal.com>