fwk: dogwood: Make 12V led follow the actual PSU state
12V led should turn on and off when we turn on and off the power supply.
Right now it hasn't because the original code was a little too lenient
and was checking downstream power supplies for its status (which is always
powered due to the 5vsb suspend mechanism).

Let's just make led follow ps_on and check in more places for this.

BUG=Easier to debug https://app.clickup.com/t/86etqaz6v
TEST=Get system to suspend, listen to clicks, led should also
change state around that same time
TEST=LED2 still follows ERS, turns on in S0, turns off in S5,
blinks when there's errors
TEST=No new 12V errors or diagnostics
BRANCH=fwk-dogwood-27111