fwk: dogwood: Fix DVT1 suspend
We used to have nice code that prevented all the 5vsb stuff
from happening in suspend, but that has since bitrotted since
we always forget to test it, so it's been broken for quite a
few versions.

Add the check back so we don't have the issue.

Let's also add more printouts so we know what's going on.

BUG=DVT1 boards cannot suspend
TEST=DVT1 boards can suspend
TEST=DVT2 boards can still suspend with force_psu_on
TEST=Newer boards suspend nicely too
BRANCH=fwk-dogwood-27111