fwk: dogwood: Disable 5vsb suspend on DVT2
On DVT2 we had the hardware problem, we thought we could rework it,
but looking back, not a lot of boards did. We do have the BIOS option
to bypass the hardware problem, but it seems like many dogfooders
forget (especially given how often they need to reset their settings
due to fan changes).

Realistically everyone on DVT2 should have this option on, anyone
with reworked boards should have moved on by now to something newer.

Let's force the power supply to stay on even on DVT2!

BUG=HW-6519, Many people forgetting to set "force power supply on"
TEST=Flash on DVT2, reset BIOS to default, board still suspends,
see EC uart for "board is old" confirmation
TEST=Try on newer boards, see power supply turn on and off still during suspend.
BRANCH=fwk-dogwood-27111