fwk: disable the hibernation timer if pressing the power button
If the system is in G3 state and the extpower is not present, EC will
enter the hibernation mode. There is a worst case that the system
de-asserts the SLP_S4 signal too late (about 4 seconds), which may cause
the user can't power on the system.

Even if we increase the hibernation timer
(4c9a9f559a4e8b76cbee948d88d05f7bf512c400), user may meet this issue.
Therefore, we should disable the hibernation timer when pressing the
power button to avoid EC triggering the hibernation and powering on
the system to fails.

BRANCH=fwk-sunflower-26784
BUG=EC enters hibernation mode, causing the system to fail to power
    on the system
TEST=Open the chassis and shutdown the system. Wait for the chipset
     state jumps to G3 and then wait for 4 seconds to press the power
     button. Check that the system can be powered on.

Signed-off-by: Josh Tsai <Josh_Tsai@compal.com>