fwk: cypd: add the new PDO profile 5V/0.9A
To ensure the battery's discharge current can meet the certification, EC
needs to reduce the type-c port output current. On the Lotus, we only limit
the output current to 1.5A, but this is not enought for the Tulip.
After discussing, EC needs to limit the port2 and port5 current to 0.9A.

This change adds the new PDO profile 5V/0.9A and initials the PDO mask
when the PD chip powers on or reset.

BRANCH=fwk-main
BUG=https://app.clickup.com/t/86ermpb5a
TEST=No logic change. Check that the Type-C port provides 3A for the first
     device while the other ports provide 1.5A.
TEST=Use cc analyzer to check the PDO is 5V/900mA if the EC sends the command
     to PD to select "PDO0 enable".

Signed-off-by: Josh Tsai <Josh_Tsai@compal.com>