fwk: Check if inputmodules are presente for diag
Instead of checking whether they're on or not.
They might be not on for different reasons right at the moment when
diagnostic checks are run.

This syncs the condition up with the LED behavior of
multifunction_leds_control() in
zephyr/program/framework/src/laptop_led.c

BRANCH=fwk-main
BUG=input_modules_reset() in enter_acpi_mode() causes the inputmodules
to be off right at the moment when diagnostics is checking them
TEST=Populate input modules and make sure diagnostic lights aren't
blinking

Signed-off-by: Daniel Schaefer <dhs@frame.work>
Co-authored-by: Josh Tsai <Josh_Tsai@compal.com>