fwk: gpu: override the GPIO flag if the chipset state is above the power domain
After reading the GPU EEPROM, EC runs the set_gpu_gpios_configuration()
and configures the gpu_3v_5v_en to output low.
When the gpu_3v_5v_en is set low, and causes the EC to be unable not
initialize the PD chip successfully.

This change add the logic in set_gpu_gpios_configuration, we could check
the GPIO configuration and if the function is high, and our power state
is above the gpio power state, then we just set the flags on instead of off

BRANCH=fwk-main
BUG=The unit cannot be charged when connecting the adapter to the rear Type-C
    port.
BUG=GPU PD chip GPU PD chip initialization fails when the EC resumes from
    hibernation.
BUG=The battery icon shows AC is active even though we don't connect the adapter
TEST=The unit can be charged when the adapter on rear type-c port
TEST=Check the EC log, there is no PD communication error when the EC
     resumes from hibernation
TEST=The battery icon display correctly if the adapter plugs in and is removed

Signed-off-by: Josh Tsai <Josh_Tsai@compal.com>