fwk: 16-inch: Monitor the fans until the diagnostic is completed
If the user opens the GPU bay and powers on the system, the fan will
not be detected. After following the instructions to install the components
and reboot the system, the screen will show "Fan not detect" message.

That is because the EC only monitors the fan speed within 8 seconds
after Diagnosing. This change monitors the fans until the diagnostic
is completed to ensure the fan can be detected after the system reboot.

BRANCH=fwk-main
BUG=https://app.clickup.com/t/86eu6r0vv
TEST=Open the GPU bay and then power on the system,
     follow the user instructions to install the components,
     short press power button to restart the system.
     The system does not show the error message "Fan not detect"

Signed-off-by: Josh Tsai <Josh_Tsai@compal.com>