fwk: cypd: check the RDO mismatch flags
If the type-c device operating current is lower than maximum operating power
in the RDO, the capability mismatch flag will be set.

This change reads the contract information after the negotiation is completed.
If the capability mismatch flag is set, call the event to check if the board
allows changing the PDO or not.

BRANCH=fwk-dogwood-27111
BUG=https://app.clickup.com/t/86etza0y1
TEST=Connect the device which will set the mismatch flag, and check
     the EC log prints "RDO Mismatch, may provide more power for this
     device"

Signed-off-by: Josh Tsai <Josh_Tsai@compal.com>