fwk: ucsi: workaround: do not bypass the invalid UCSI data to host
PD chip will set the negotiated_power_level_change when the port
is disconnected. But the UCSI specification describes that this
bit shall be set by the PPM whenever a Power contract is established
or renegotiated.

This is a workaround to clear the invalid data format to pass the
HLK test item: "UCSI Get Connector Status - Negotiated Power Level
Change [Type-C MUTT]"

BRANCH=fwk-main
BUG=https://app.clickup.com/t/86etqf4jh
BUG=https://app.clickup.com/t/86er6n76t
BUG=https://app.clickup.com/t/86etjdkpx
TEST=Run the HLK test item "UCSI Get Connector Status - Negotiated
     Power Level Change [Type-C MUTT]" can get the pass result with
     all type-c ports

Signed-off-by: Josh Tsai <Josh_Tsai@compal.com>