fwk: raa489300: should not enter low power PTM mode if the charger is in buck mode
In the charger specification, there is no description that the low power
mode only can be updated when the charger is in PTM mode.

But to avoid the low power PTM mode entering and exiting more time during
shutdown the system, this change add the condition to avoid the charger
updating the low power PTM mode bit when the charger is in buck mode.

BRANCH=fwk-tulip-29196
BUG=https://app.clickup.com/t/86etjj91x
TEST=Use the ec console command "i2ctrace enable 3 0x4a" to check
     the low power PTM mode only be updated when the charger is in
     the PTM mode.

Signed-off-by: Josh Tsai <Josh_Tsai@compal.com>