fwk: dogwood: Repurpose "APU_EC_INT" to be a 4th RAM ID bit
We're running out of RAM ID pins going to the AP. This is a problem because
we need a little more RAM options than expected (eg: multi-sourcing).

Since we can't really change the board anymore to add more bits,
we need another method.

Luckly the APU_EC_INT signal going between the EC->AP is unused
(despite its name). It was originally added for some HID interrupts.
We never needed that.
This signal is also not to be confused with the normal EC->AP interrupt
method that happens inband over eSPI,

The proposal is for the AP to use this APU_EC_INT as a 4th RAM ID bit.

How will the EC know when to assert this? We'll use the EC board ID
(toggling between 9 and 10) to disambiguate between the 2 states
for this 4th RAM_ID bit.
NOTE: otherwise board_id 9 and 10 are exactly the same as far as
the EC is concerned.

BRANCH=fwk-dogwood-27111
BUG=https://app.clickup.com/t/86euhhg7k
See https://docs.google.com/presentation/d/1n1WNI4WoaVsPbWK_gPJw2rZpELAfXCdn for more details.
TEST=on dogwood, when board version id = 10, APU_EC_INT will keep low.