fwk: ensure the charger maximum voltage is battery->voltage_normal
If the user hot-plugs the battery, the EC code does not clear the
battery information. It causes the charger manager to write the
battery->voltage into the charger maximum voltage instead of
battery->voltage_normal.

This change clears the battery voltage if the battery is disconnected
to ensure the charger's maximum voltage value is the same as the dead battery.

BRANCH=fwk-main
BUG=Charger maximum voltage value is different between the dead battery
    mode and hot-plugging the battery.
TEST=Connect the adapter to power up the EC, check the charger register
    0x15 value is 0x3c28 (15400 mV, 55W battery voltage_normal)
TEST=Connect the battery, the charger register 0x15 value is 0x4588
     (17800 mV, 61W battery voltage_max)
TEST=Remove the battery, the charger register 0x15 value is 0x3c78
     (15480 mV, 61W battery voltage_max)

Signed-off-by: Josh Tsai <Josh_Tsai@compal.com>