dsp_comms: Add client retry
In some cases it was found that the EC is suspended when the ISH tries
to send an I2C message. This leaves the EC nacking the transaction.
Adding a simple retry based on Kconfig values. We set the default
retry to 30 times to make 100% sure the message goes through, though
in tests we've only seen a small number of retries.

BUG=b:433638452, b:440698037
TEST=Suspend DUT, wait for EC to suspend, change to tablet mode

Change-Id: I345929f7de128f2122efbe77ce1cd0a5fc5f4e4b
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/ec/+/6921192
Commit-Queue: Yuval Peress <peress@google.com>
Reviewed-by: Li1 Feng <li1.feng@intel.com>
Reviewed-by: Deepti Deshatty <deepti.deshatty@intel.com>
Tested-by: Yuval Peress <peress@google.com>
Reviewed-by: Gwendal Grignou <gwendal@chromium.org>
Auto-Submit: Yuval Peress <peress@google.com>