uldrenite: Adjust the priority of gmr_tablet_switch()
There is an issue where the hotkey cutoff fails.
This is because the execution priority of gmr_tablet_switch_init() is too high, causing it to run before the project's own initialization process.
As a result, it incorrectly reads the status of an unused pin (tablet_mode_l) and locks the keyboard.

This patch adjusts the execution priority to prevent this issue.

BUG=b:446605593
TEST=cuf-off with hotkey success.

Change-Id: Ic8528fc72335f24f39fc81f3c213a694cd21ba46
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/ec/+/6975695
Reviewed-by: Logan Liao <logan_liao@compal.corp-partner.google.com>
Code-Coverage: Zoss <zoss-cl-coverage@prod.google.com>
Reviewed-by: Scott Collyer <scollyer@chromium.org>
Reviewed-by: Allan Almiranez <allanalmiranez@google.com>
Reviewed-by: Elthan Huang <elthan_huang@compal.corp-partner.google.com>
Tested-by: Logan Liao <logan_liao@compal.corp-partner.google.com>
Reviewed-by: Yuval Peress <peress@google.com>
Commit-Queue: Logan Liao <logan_liao@compal.corp-partner.google.com>