util/update_release_branch.py: Fix incorrect fpmcu directory path
The path to the fpmcu directory in the script appears to have been
incorrect due to a typo. The directory itself has existed since the
the beginning of the project ( http://shortn/_IToDMhBBTj ). This change
corrects the path to the existing directory. The incorrect path due to
the typo is not believed to have caused any issues.

BUG=b:444087489
TEST=None

Change-Id: I299d86e44c0d7bf17e361ba192caed42a48e65be
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/ec/+/6997967
Reviewed-by: Tom Hughes <tomhughes@chromium.org>
Tested-by: Firas Sammoura <fsammoura@google.com>
Commit-Queue: Firas Sammoura <fsammoura@google.com>
Code-Coverage: Zoss <zoss-cl-coverage@prod.google.com>