[owners] Remove jettrink@chromium.org from board/c2d2/OWNERS
This suggested change is automatically generated based on group
memberships and affiliations.

Please approve this change and vote the highest CR. This will keep the
OWNERS file tidy. We ask that you do not ignore this change and either
approve it if the OWNER is no longer active or deny it if the OWNER is
continuing to review CLs (including under a different affiliation).
This CL can always be reverted if the understanding changes.

If this change is in error, vote the lowest CR value (i.e. reject the
CL) and the bot will abandon it.

See the owner's recent review activity for context:
https://chromium-review.googlesource.com/q/jettrink@chromium.org

To report an issue, file a bug in the Infra>Codereview component.

Change-Id: Ib39b989c59e6a0e86d4e1c4f7f72a4f1ba13a1ec
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/ec/+/7017599
Commit-Queue: Michał Headcrab Barnaś <barnas@google.com>
Code-Coverage: Zoss <zoss-cl-coverage@prod.google.com>
Auto-Submit: Owners Cleanup <swarming-tasks@owners-cleanup-prod.google.com.iam.gserviceaccount.com>
Reviewed-by: Keith Short <keithshort@chromium.org>
Tested-by: Michał Headcrab Barnaś <barnas@google.com>