fwk: sakura: delete the unused-pins node
Used the gpioget to check that the EC gpio status and found the
"gpio_ec_amp_pwr_en" and "gpio_ec_ts_pwr_en" status are incorrect
(expected low but got high).
This change puts the "/delete-node/ unused-pins" to the correct
location to ensure the gpio status is correct.

BRANCH=fwk-main
BUG=ec_amp_pwr_en and ec_ts_pwr_en cannot drive low
TEST=use gpioget to check that ec_amp_pwr_en and ec_ts_pwr_en are
     low when the chipset state is at S5, and are high when the
     chipset state is at S0.

Signed-off-by: Josh Tsai <Josh_Tsai@compal.com>