crisis_rcvry_util.py: Increase retry count
Be more resilient to spurious CRC errors when flashing the microchip EC.

Getting about 10 "CRC mismatch" errors when flashing the EC on
satlab-*-ptl-rack1-bay27 (i.e. RVPs deployed in the lab). This seems to
be less of an issue when the DuT/servo is connected directly to a
laptop/workstation. In any case, the original retry count of 3 is not
practical. Let's try 100 since we really want best effort here and it's
low enough that we don't get stuck trying to flash in an eventual
failure case.

BUG=none
TEST=flashed microchip EC on PTL RVP in satlab

Change-Id: I4b0ee16ea3e9b1d9b2b95858176d9f80d8345084
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/ec/+/7030012
Reviewed-by: Scott Collyer <scollyer@chromium.org>
Tested-by: caveh jalali <caveh@chromium.org>
Code-Coverage: Zoss <zoss-cl-coverage@prod.google.com>
Commit-Queue: caveh jalali <caveh@chromium.org>