fwk: cypd: rename CONFIG_PD_CCG6_WAIT_STABLE_TIMER
In the commit fe37b179c8cada0b80c803f88f180e991ab3c7ae, we added the
delay time to wait for the CCG6 PD to power on ready. But the wait
for the PD stable feature is not for the CCG6 only.

This change renames the CONFIG_PD_CCG6_WAIT_STABLE_TIMER and removes
"depends on PD_CHIP_CCG6" in the Kconfig file.

BRANCH=fwk-main
BUG=None
TEST=No logic change

Signed-off-by: Josh Tsai <Josh_Tsai@compal.com>