fwk: sakura: delay 20ms to wait for the PD chip power on ready.
According to Weikeng's recommandation, EC shouldn't communicate
with PD before initializing ready. On the Sakura unit, if EC
communicates with the PD chip before it is initially ready will
cause that the PD lockes the SCL to low.

This change adds the delay time 20ms to wait for the PD ready to
ensure the EC communicates the PD chip after the PD asserts the
interrupt to notify the EC to check the status.

BRANCH=fwk-main
BUG=https://app.clickup.com/t/86ev202w1
TEST=use the logical analyzer to check the EC communicates with
     PD chip after the PD asserts the interrupt.

Signed-off-by: Josh Tsai <Josh_Tsai@compal.com>