power/qcom_exp: Reset PMIC passthru signals on power-on
The passthru signals from the EC to the PMIC were not explicitly
reset during the power-on sequence. This could lead to a stale state,
causing the AP to boot with an incorrect power-on reason or due to an
unintended flow.

Update the power-on flow to actively reset these signals. This change
guarantees that the AP starts with a proper power-on reason and
boots only with the intended flow.

BUG=b:445857032
TEST=zmake build %bluey
TEST=able to get the correct power on reason
LOW_COVERAGE_REASON=b:422080009 tracking test code

Change-Id: I4aa1329e5fc70dbef89ceb65671926606d5ac307
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/ec/+/7037477
Reviewed-by: Wai-Hong Tam <waihong@google.com>
Tested-by: Jayvik Desai <jayvik@google.com>
Auto-Submit: Jayvik Desai <jayvik@google.com>
Commit-Queue: Jayvik Desai <jayvik@google.com>
Code-Coverage: Zoss <zoss-cl-coverage@prod.google.com>