fwk: ucsi: don't resend the ucsi command if not enable the notification
EC will crash if the user plugs or unplugs a type-c device in the
standby mode. The root cause is that EC will point to the wrong address
if the first PD chip state is not powered.

This change disables resend_ucsi_connector_change_event() function to
avoid the wrong ucsi event being processed with the pure OS (no UCSI driver)

I don't fix the wrong pointer issue with this change. This condition only
for a workaround; the UCSI driver will be yellow bang if one PD chip is not
powered on. So this issue does not happen on PD0 and PD1 are ready.

BRANCH=fwk-main
BUG=wrong pointer causes the EC crash
TEST=plug or unplug a device in standby mode with pure OS, the EC
     does not crash.

Signed-off-by: Josh Tsai <Josh_Tsai@compal.com>
(cherry picked from commit fc05c40bf55b9caa5dd2c02bcf90c5bb73c0cc02)