fwk: sakura: enable PD chip0
We found an issue that when EC sends the reset command to the PD
chips, the PD chips lock the I2C clock and fail to communicate.

In this change 76bda79df9af29e3c0d7bd43539d7e88c7635a6f, we disabled
to communicate with the right side PD chip to ensure EC does not
trigger the watchdog.

Now we plan to revert the change, because this issue only happens
once the EC sends reset command to PD chip. It will not impact the
user.

BRANCH=fwk-main
BUG=https://app.clickup.com/t/86ev203vt
TEST=Can initailize the right side PD chip
TEST=plug in the device can receive the interrupt event

Signed-off-by: Josh Tsai <Josh_Tsai@compal.com>