fwk: rename update_pmf_events to update_cpu_power_limit_events
In the past project, only AMD platform uses the CCG8 and supports the
EPR mode, so we named the function update_pmf_events().
Now, we use the CCG8 and support the EPR mode on the Intel platform.

This change renames the update_pmf_events to update_cpu_power_limit_events.

BRANCH=fwk-main
BUG=None
TEST=no logic change

Signed-off-by: Josh Tsai <Josh_Tsai@compal.com>