util: use lowercase for min/max/clamp
Change the min/max/clamp utilities in the legacy firmware to lowercase.
The convention for both Linux and Zephyr is that min/max/clamp are
implemented as inline function and single evaluation version, while MIN
MAX and CLAMP are trivial implementation and can be used in macros.

For non Zephyr project this is a no-op, but aligns the code with the
Linux and now Zephyr convention.

For Zephyr projects it means that shimmed drivers now use the single
evaluation version, as they were originally developed, which beside the
performance and code size (marginal) benefits, is safe from potential
race conditions due to the double evaluation nature of the Zephyr
MIN/MAX/CLAMP versions.

BUG=none
TEST=cq dry run

Change-Id: Ia7e49ef388692a4f8a6724a929eecf4f6c9f728e
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/ec/+/7080810
Reviewed-by: Keith Short <keithshort@chromium.org>
Reviewed-by: Firas Sammoura <fsammoura@google.com>
Tested-by: Fabio Baltieri <fabiobaltieri@google.com>
Commit-Queue: Fabio Baltieri <fabiobaltieri@google.com>