fwk: cypd: EC should wait for tBootWait time after resetting the PD
If the EC issues a reset command to the PD chip,
the PD will lock the I2C Bus. This prevents the EC from communicating
with the PD chip and consequently triggers the watchdog.

According to Weikeng's analysis, the PD chip enters the boot mode after
being reset. The EC is required to wait for tBootWait period after receiving
the boot mode interrupt.

However, the tBootWait time is not a fixed value across different PD
chips. The EC should wait for the maximum possible value to ensure this
issue does not occur.

BRANCH=fwk-main
BUG=https://app.clickup.com/t/86ev203vt
TEST=Use the command "cypdctl reset" to reset the PD chip. EC
     can initialize the PD chip completely. The I2C bus will not be locked.

Signed-off-by: Josh Tsai <Josh_Tsai@compal.com>