fwk: CCG6: set the active current limit before multi-port switching
When the EC processes a multi-port switch, the sink path is enabled
first, and then the input current limit for the active port is set.
This sequence can cause a lower power adapter to operate with a higher
input current limit , potentially triggering an OCP event.

To ensure the adapter operates within a safe operating current, the EC
should set the input current limit to the active one before enabling the
sink path or completing the multi-port switching.

BRANCH=fwk-main
BUG=https://app.clickup.com/t/86ev81hyn
TEST=The EC log will not print the AC off during the multi-port switching

Signed-off-by: Josh Tsai <Josh_Tsai@compal.com>