keyboard_8042_fn: process Fn key if no other function keys are pressed
Add a condition to send a press and release of the Fn button if it gets
pressed and released without pressing any other function keys in the
meantime. This ensures that the button can be used to wakeup the device.

BUG=b:446679629
TEST=test config on a rex board

Change-Id: I2454f8b048b1da07933413aa48a8eca5926d7de0
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/ec/+/7119158
Tested-by: Fabio Baltieri <fabiobaltieri@google.com>
Commit-Queue: Fabio Baltieri <fabiobaltieri@google.com>
Reviewed-by: caveh jalali <caveh@chromium.org>
Code-Coverage: Zoss <zoss-cl-coverage@prod.google.com>
Reviewed-by: Abe Levkoy <alevkoy@chromium.org>